baseline_model/requirements.txt
backcall>=0.1.0
certifi>=2018.11.29
chardet>=3.0.4
cloudpickle>=0.6.1
cycler>=0.10.0
cytoolz>=0.9.0.1
dask>=1.0.0
decorator>=4.3.0
idna>=2.8
imageio>=2.4.1
ipykernel>=5.1.0
ipython>=7.2.0
ipython-genutils>=0.2.0
jedi>=0.13.2
jupyter-client>=5.2.4
jupyter-core>=4.4.0
kiwisolver>=1.0.1
matplotlib>=3.0.2
# mkl-fft>=1.0.6  # Failed to install, maybe these packages have been removed
# mkl-random>=1.0.2
networkx>=2.2
numpy>=1.15.4
olefile>=0.46
parso>=0.3.1
pexpect>=4.6.0
pickleshare>=0.7.5
Pillow>=5.3.0
prompt-toolkit>=2.0.7
ptyprocess>=0.6.0
Pygments>=2.3.0
pyparsing>=2.3.0
pyreadline>=2.1
python-dateutil>=2.7.5
pytz>=2018.7
PyWavelets>=1.0.1
pyzmq>=17.1.2
requests>=2.21.0
scikit-image>=0.14.1
scipy>=1.1.0
six>=1.12.0
toolz>=0.9.0
torch>=1.0.0
torchvision>=0.2.1
tornado>=5.1.1
traitlets>=4.3.2
urllib3>=1.24.1
vk>=2.0.2
wcwidth>=0.1.7
